+++ src/components/AddContent.vue
@@ -1,27 +1,67 @@
 <template>
-  <div>
-    <h4>Add {{ type }}</h4>
+  <q-modal ref="addContentModal" v-model="showAddContent" content-classes="add-content-modal">
+    <q-btn
+      color="primary"
+      @click.native="showAddContent = false"
+      icon="fas fa-times"
+      class="float-right"
+      size="sm"
+    />
+    <h4>Add {{ readableType }}</h4>
     <div class="row gutter-sm">
       <div class="col-12">
         <q-input type="text" float-label="Title" v-model="title" />
       </div>
+      <div class="col-12" v-if="type === 'osermon' || type === 'olesson'">
+        <h5>Choose a template...</h5>
+      </div>
+      <div class="col-xs-12 col-md-6" v-if="type === 'osermon' || type === 'olesson'">
+        <q-card>
+          <q-card-title>Expositional</q-card-title>
+          <q-card-main>
+            <p>Start out with a passage and break it down, verse-by-verse!</p>
+            <q-input type="text" float-label="Bible Passage" v-model="bibleRef" />
+            <br/>
+            <br/>
+            <q-btn color="primary" class="float-right" :disabled="bibleRef === ''" @click.native="add('expo')">Choose</q-btn>
+          </q-card-main>
+        </q-card>
+      </div>
+      <div class="col-xs-12 col-md-6" v-if="type === 'osermon' || type === 'olesson'">
+        <q-card>
+          <q-card-title>3-Point</q-card-title>
+          <q-card-main>
+            <p>The familiar 3-point {{ type }} to get you started!</p>
+            <q-btn color="primary" class="float-right" @click.native="add('3point')">Choose</q-btn>
+          </q-card-main>
+        </q-card>
+      </div>
       <div class="col-12">
-        <q-btn color="primary" class="float-right" @click.native="add">Add {{ type }}</q-btn>
+        <h5 v-if="type === 'osermon' || type === 'olesson'">...Or start from scratch</h5>
+        <q-btn color="primary" class="float-right" @click.native="add('blank')">Add Blank {{ readableType }}</q-btn>
       </div>
     </div>
-  </div>
+  </q-modal>
 </template>
 
 <script>
-import { Notify } from 'quasar'
+import { Notify, format } from 'quasar'
+const { capitalize } = format
 
 export default {
-  props: ['type', 'modalFin'],
+  props: ['type'],
   // name: 'ComponentName',
   data () {
     return {
+      showAddContent: false,
       title: '',
-      types: ['rseries', 'oseries', 'olesson', 'osermon', 'oscratch']
+      types: ['rseries', 'oseries', 'olesson', 'osermon', 'oscratch'],
+      bibleRef: ''
+    }
+  },
+  computed: {
+    readableType: function () {
+      return capitalize(this.type.slice(1))
     }
   },
   mounted () {
@@ -31,13 +71,15 @@ export default {
     init () {
       this.title = ''
     },
-    add () {
+    add (template) {
       if (this.types.includes(this.type)) {
         var obj = {
-          title: this.title
+          title: this.title,
+          template: template,
+          prefs: this.$root.user.prefs[this.type + 'Structure']
         }
         this.$database.add(this.type, obj, (res) => {
-          this.modalFin()
+          this.showAddContent = false
           Notify.create({
             message: this.type + ' created!',
             type: 'positive',
@@ -46,6 +88,9 @@ export default {
           this.$router.push({ name: this.type, params: { id: res._id } })
         })
       }
+    },
+    show () {
+      this.showAddContent = true
     }
   }
 }