+++ src/components/modules/Activity.vue
@@ -3,7 +3,7 @@
     <div v-show="!data.editing || data.editing !== $firebase.auth.currentUser.uid">
       <q-card-title>
         <q-icon v-show="!data.editing" class="float-right cursor-pointer" name="fas fa-ellipsis-v" color="primary" size="1rem">
-          <q-popover>
+          <q-popover anchor="bottom right" self="top right">
             <q-list>
               <q-item link v-close-overlay @click.native="edit(id)">Edit</q-item>
               <q-item link @click.native="remove(id)">Delete</q-item>
@@ -23,13 +23,13 @@
         <q-icon link class="float-right cursor-pointer" name="fas fa-times" size="1rem" @click.native="close" />
         <div class="row gutter-sm">
           <div class="col-12">
-            <q-input v-model="data.title" float-label="Title" />
+            <q-input v-model="data.title" float-label="Title" autofocus />
           </div>
           <div class="col-12">
             <q-input type="number" v-model="data.time" float-label="Estimated Time (in minutes)" />
           </div>
           <div class="col-12">
-            <editor :text.sync="data.text" />
+            <editor :text.sync="data.text" :save="autoSave" />
           </div>
           <div class="col-12">
             <q-btn color="primary" @click.native="save(id)">Save</q-btn>
@@ -52,6 +52,14 @@ export default {
   props: [ 'id', 'data', 'edit', 'save', 'close', 'remove' ],
   data () {
     return {}
+  },
+  methods: {
+    autoSave () {
+      console.log('auto save text')
+      this.$firebase.ref(this.$parent.$parent.type, 'modules', this.$parent.$parent.id).child(this.id).update({
+        text: this.data.text
+      })
+    }
   }
 }
 </script>