+++ src/components/modules/Question.vue
@@ -3,7 +3,7 @@
     <div v-show="!data.editing || data.editing !== $firebase.auth.currentUser.uid">
       <q-card-title>
         <q-icon v-show="!data.editing" class="float-right cursor-pointer" name="fas fa-ellipsis-v" color="primary" size="1rem">
-          <q-popover>
+          <q-popover anchor="bottom right" self="top right">
             <q-list>
               <q-item link v-close-overlay @click.native="edit(id)">Edit</q-item>
               <q-item link @click.native="remove(id)">Delete</q-item>
@@ -20,13 +20,10 @@
         <q-icon link class="float-right cursor-pointer" name="fas fa-times" size="1rem" @click.native="close" />
         <div class="row gutter-sm">
           <div class="col-12">
-            <q-input v-model="data.title" float-label="Title" />
+            <q-input v-model="data.text" float-label="Question" type="textarea" :max-height="100" :min-rows="1" autofocus />
           </div>
           <div class="col-12">
             <q-input type="number" v-model="data.time" float-label="Estimated Time (in minutes)" />
-          </div>
-          <div class="col-12">
-            <q-input v-model="data.text" float-label="Question" type="textarea" :max-height="100" :min-rows="1" />
           </div>
           <div class="col-12">
             <q-btn color="primary" @click.native="save(id)">Save</q-btn>