+++ src/components/modules/Section.vue
@@ -3,7 +3,7 @@
     <div v-show="!data.editing || data.editing !== $firebase.auth.currentUser.uid">
       <q-card-title>
         <q-icon v-show="!data.editing" class="float-right cursor-pointer" name="fas fa-ellipsis-v" size="1rem">
-          <q-popover>
+          <q-popover anchor="bottom right" self="top right">
             <q-list>
               <q-item link v-close-overlay @click.native="edit(id)">Edit</q-item>
               <q-item link @click.native="remove(id)">Delete</q-item>
@@ -19,7 +19,7 @@
         <q-icon link class="float-right cursor-pointer" name="fas fa-times" size="1rem" @click.native="close" />
         <div class="row gutter-sm">
           <div class="col-12">
-            <q-input v-model="data.title" float-label="Title" @keyup.enter="save(id)" />
+            <q-input v-model="data.title" float-label="Title" @keyup.enter="save(id)" class="dark-label" autofocus />
           </div>
           <div class="col-12">
             <q-btn color="primary" @click.native="save(id)">Save</q-btn>
@@ -42,4 +42,9 @@ export default {
 </script>
 
 <style>
+
+.dark-label .q-if-label {
+  color: white !important;
+}
+
 </style>